fixed invisible bullets bug

+++ js/main.js
@@ -8,21 +8,16 @@ export const GAME_TIMER_INTERVAL = 50; // задаёт интервал врем
 export const PLAYER_LIFE_COUNT = 3;
 export const ENEMY_TANKS_COUNT = 21;
 export const CELL_SIZE = 64;
-export const BULLET_SIZE = 8;
-export const IS_GAME_OVER = false;
+export const BULLET_SIZE = 1;
 
 startGame();
 
 function startGame() {
+    let isGameOver = false;
+    
     document.querySelector("main").innerHTML = document.getElementById("game-template").innerHTML;
-
     const $playerLives = document.getElementById("player-lives");
     const $enemyLives = document.getElementById("enemy-lives")
-    const intervalObject = {value: GAME_TIMER_INTERVAL};
-    window.intervalObject = intervalObject;
-
-    window.livesCounter = new LivesCounter($playerLives, $enemyLives, PLAYER_LIFE_COUNT, ENEMY_TANKS_COUNT, startGame);
-
     const $gameMap = document.getElementById("game-map");
 
     let updatableElements = [];
@@ -33,6 +28,7 @@ function startGame() {
     window.updatableElements = updatableElements;
     window.walls = walls;
     window.enemyTanks = enemyTanks;
+    window.livesCounter = new LivesCounter($playerLives, $enemyLives, PLAYER_LIFE_COUNT, ENEMY_TANKS_COUNT, restartGame);
 
     gameInitialization();
 
@@ -93,59 +89,57 @@ function startGame() {
         });
     }
     function gameLoop() {
-        if (IS_GAME_OVER !== true) {
-
-            /**
-             * вот именно в функции gameStep стоит разместить код, который будет выполняться на каждом шаге игрового цикла
-             */
+        if (isGameOver !== true) {
             gameStep();
-
-
+            
             setTimeout(function () {
                 gameLoop()
-            }, intervalObject.value);
+            }, 50);
         }
     }
     function gameStep() {
-        checkCollisions();
+        checkCollisions(enemyTanks, playerTank, walls);
         updatableElements.forEach(element => element.update());
     }
-
-    function checkCollisions() {
-        const tanks = [...enemyTanks, playerTank];
-
-        for (let tank of tanks) {
-            checkTankWallsCollisions(tank);
-            checkBulletWallsCollisions(tank.bullet);
-        }
-
-        checkTanksCollision(tanks);
-        checkEnemyBulletsPlayerTankCollision();
-        if(playerTank.bullet.isFlying) checkPlayerBulletEnemyTanksCollision();
+    function restartGame() {
+        isGameOver = true;
+        startGame();
     }
 
-    function checkTankWallsCollisions(tank) {
-        tank.availableDirections = {LEFT: true, RIGHT: true, TOP: true, BOTTOM: true};
+    function checkCollisions(enemyTanks, playerTank, walls) {
+        const tanks = [...enemyTanks, playerTank];
 
-        if (tank.x <= 0) tank.availableDirections.LEFT = false;
-        if (tank.x >= MAP_WIDTH - 1) tank.availableDirections.RIGHT = false;
-        if (tank.y <= 0) tank.availableDirections.TOP = false;
-        if (tank.y >= MAP_HEIGHT - 1) tank.availableDirections.BOTTOM = false;
+        
+        checkTankWallsCollisions(tanks, walls);
+        checkBulletWallsCollisions(tanks, walls);
+        checkTanksCollision(tanks);
+        checkEnemyBulletsPlayerTankCollision(enemyTanks, playerTank);
+        if(playerTank.bullet.isFlying) checkPlayerBulletEnemyTanksCollision(enemyTanks, playerTank);
+    }
+    function checkTankWallsCollisions(tanks, walls) {
+        for(let tank of tanks) {
+            tank.availableDirections = {LEFT: true, RIGHT: true, TOP: true, BOTTOM: true};
 
-        for (let wall of walls) {
-            let dx = tank.x - wall.x;
-            let dy = tank.y - wall.y;
-            if ((dx <= 1 && dx > 0) && Math.abs(dy) < 1) {
-                tank.availableDirections.LEFT = false;
-            }
-            if ((dx >= -1 && dx < 0) && Math.abs(dy) < 1) {
-                tank.availableDirections.RIGHT = false;
-            }
-            if ((dy <= 1 && dy > 0) && Math.abs(dx) < 1) {
-                tank.availableDirections.TOP = false;
-            }
-            if ((dy >= -1 && dy < 0) && Math.abs(dx) < 1) {
-                tank.availableDirections.BOTTOM = false;
+            if (tank.x <= 0) tank.availableDirections.LEFT = false;
+            if (tank.x >= MAP_WIDTH - 1) tank.availableDirections.RIGHT = false;
+            if (tank.y <= 0) tank.availableDirections.TOP = false;
+            if (tank.y >= MAP_HEIGHT - 1) tank.availableDirections.BOTTOM = false;
+
+            for (let wall of walls) {
+                let dx = tank.x - wall.x;
+                let dy = tank.y - wall.y;
+                if ((dx <= 1 && dx > 0) && Math.abs(dy) < 1) {
+                    tank.availableDirections.LEFT = false;
+                }
+                if ((dx >= -1 && dx < 0) && Math.abs(dy) < 1) {
+                    tank.availableDirections.RIGHT = false;
+                }
+                if ((dy <= 1 && dy > 0) && Math.abs(dx) < 1) {
+                    tank.availableDirections.TOP = false;
+                }
+                if ((dy >= -1 && dy < 0) && Math.abs(dx) < 1) {
+                    tank.availableDirections.BOTTOM = false;
+                }
             }
         }
     }
@@ -175,21 +169,24 @@ function startGame() {
             }
         }
     }
-    function checkBulletWallsCollisions(bullet) {
-        const isBulletOutsideMap = bullet.x > MAP_WIDTH || bullet.x < 0 || bullet.y < 0 || bullet.y > MAP_HEIGHT;
-        if(isBulletOutsideMap) {
-            bullet.isFlying = false;
-            return;
-        }
-
-        for(let wall of walls) {
-            if(isSquareInsideSquare(wall.x, wall.y, CELL_SIZE, bullet.x, bullet.y, BULLET_SIZE)) {
-                wall.destroy();
+    function checkBulletWallsCollisions(tanks, walls) {
+        for(let tank of tanks) {
+            const bullet = tank.bullet;
+            const isBulletOutsideMap = bullet.x > MAP_WIDTH || bullet.x < 0 || bullet.y < 0 || bullet.y > MAP_HEIGHT;
+            if(isBulletOutsideMap) {
                 bullet.isFlying = false;
+                return;
+            }
+
+            for(let wall of walls) {
+                if(isSquareInsideSquare(wall.x, wall.y, CELL_SIZE, bullet.x, bullet.y, BULLET_SIZE)) {
+                    wall.destroy();
+                    bullet.isFlying = false;
+                }
             }
         }
     }
-    function checkPlayerBulletEnemyTanksCollision() {
+    function checkPlayerBulletEnemyTanksCollision(enemyTanks, playerTank) {
         for(let enemyTank of enemyTanks) {
             if(isSquareInsideSquare(enemyTank.x, enemyTank.y, CELL_SIZE, playerTank.bullet.x, playerTank.bullet.y, BULLET_SIZE)) {
                 enemyTank.kill();
@@ -197,8 +194,9 @@ function startGame() {
             }
         }
     }
-    function checkEnemyBulletsPlayerTankCollision() {
+    function checkEnemyBulletsPlayerTankCollision(enemyTanks, playerTank) {
         for(let enemyTank of enemyTanks) {
+            if(!enemyTank.bullet.isFlying) return;
             if(isSquareInsideSquare(playerTank.x, playerTank.y, CELL_SIZE, enemyTank.bullet.x, enemyTank.bullet.y, BULLET_SIZE)) {
                 playerTank.kill();
                 enemyTank.bullet.isFlying = false;
